Refactor dashboard options
